improved requests

+++ Express/client/performRequest.js
@@ -1,6 +1,6 @@
 import { getToken } from "./setupLogin";
 
-export const setupPerformRequest = async (element) => {
+export const setupPerformRequest = (element, onGetMoviesCb) => {
   element.addEventListener("click", async () => {
     try {
       const response = await fetch("./api", {
@@ -15,16 +15,23 @@ export const setupPerformRequest = async (element) => {
       }
       const data = await response.json();
       console.log(data);
+      if (typeof onGetMoviesCb === "function") {
+        onGetMoviesCb(data);
+      }
     } catch (err) {
       console.log(err);
     }
   });
 };
 
-export const setupPerformDelete = async (element) => {
-  element.addEventListener("click", async () => {
+export const setupPerformDelete = (element, onDeleteMovieCb) => {
+  element.addEventListener("click", async (e) => {
+    if (!e.target.classList.contains("delete-movie-btn")) {
+      return;
+    }
+    const id = e.target.closest(".movie-card").id;
     try {
-      const response = await fetch(`./api/2`, {
+      const response = await fetch(`./api/${id}`, {
         method: "DELETE",
         headers: {
           "Content-Type": "application/json",
@@ -34,8 +41,9 @@ export const setupPerformDelete = async (element) => {
       if (response.status !== 200) {
         throw new Error(response.statusText);
       }
-      const data = await response.json();
-      console.log(data);
+      if (typeof onDeleteMovieCb === "function") {
+        onDeleteMovieCb();
+      }
     } catch (err) {
       console.log(err);
     }
